+++ api/src/routes.js
@@ -1,4 +1,3 @@
-const { on, setPreflightHeaders } = require('./router');
 
 require('./routes/auth');
 require('./routes/config');
@@ -7,5 +6,3 @@ require('./routes/badges');
 require('./routes/system');
 require('./routes/icons');
 require('./routes/version');
-
-on('OPTIONS', '*', (_, res) => { setPreflightHeaders(res); res.writeHead(204); res.end(); });